+++ source/util.js
@@ -1,5 +1,8 @@
-// Checks whether the entire input sequence can be matched
-// against the regular expression.
+/**
+ * Checks whether the entire input sequence can be matched
+ * against the regular expression.
+ * @return {boolean}
+ */
 export function matchesEntirely(text, regular_expression) {
 	// If assigning the `''` default value is moved to the arguments above,
 	// code coverage would decrease for some weird reason.