+++ CHANGELOG.md
@@ -53,10 +53,17 @@ And edit the README:
 
 <!-- (breaking change) Changed `countries` and `country_calling_codes` properties in metadata: now they're not properties but rather elements of an array (`countries` is an array now rather than an object; `countries` is `metadata[0]` and `country_calling_codes` is `metadata[1]`). If you were using a custom-generated metadata then it has to be re-generated for the new version. -->
 
+1.7.38 / 04.02.2020
+===================
+
+* Removed the `"001"` country code ("Non-Geographic Entity"): now in case of "non-geographic" phone numbers their `country` is just `undefined`. Instead, `PhoneNumber` class has an `.isNonGeographic()` method.
+
+* Fixed "non-geographic" numbers `.isPossible() === false` bug.
+
 1.7.35 / 03.02.2020
 ===================
 
-* Fixed "Non-Geographical Entities" (`001` country code).
+* Fixed "Non-Geographic Entities" (`001` country code).
 
 1.7.32 / 03.02.2020
 ===================